Relocate telnetconsole to extensions/
