+++ scrapy/templates/project/module/settings.py.tmpl
@@ -56,7 +56,7 @@ NEWSPIDER_MODULE = '$project_name.spiders'
 # Enable or disable extensions
 # See http://scrapy.readthedocs.org/en/latest/topics/extensions.html
 #EXTENSIONS = {
-#    'scrapy.telnet.TelnetConsole': None,
+#    'scrapy.extensions.telnet.TelnetConsole': None,
 #}
 
 # Configure item pipelines
